include code that actually throws the errors
this readme was confusing without including the code that actually causes the errors to be thrown. I know it's in the file itself, but it's worth adding the lines back into these code examples. They are discussed in the readme, I don't see why they're not included here (it's just 3 lines after all)